+++ android/src/main/java/com/capacitorjs/community/plugins/bluetoothle/Device.kt
@@ -17,6 +17,7 @@ import android.content.Intent
 import android.content.IntentFilter
 import android.os.Build
 import android.os.Handler
+import android.os.HandlerThread
 import android.os.Looper
 import androidx.annotation.RequiresApi
 import com.getcapacitor.Logger
@@ -167,6 +168,24 @@ class Device(
     private var currentMtu = -1
 
     private final var UUID_PRINTER = UUID.fromString("00001101-0000-1000-8000-00805F9B34FB");
+    private lateinit var callbacksHandlerThread: HandlerThread
+    private lateinit var callbacksHandler: Handler
+
+    private fun initializeCallbacksHandlerThread() {
+        synchronized(this) {
+            callbacksHandlerThread = HandlerThread("Callbacks thread")
+            callbacksHandlerThread.start()
+            callbacksHandler = Handler(callbacksHandlerThread.looper)
+        }
+    }
+
+    private fun cleanupCallbacksHandlerThread() {
+        synchronized(this) {
+            if (::callbacksHandlerThread.isInitialized) {
+                callbacksHandlerThread.quitSafely()
+            }
+        }
+    }
 
     private val gattCallback: BluetoothGattCallback = object : BluetoothGattCallback() {
         override fun onConnectionStateChange(
@@ -186,6 +205,7 @@ class Device(
                 bluetoothGatt?.close()
                 bluetoothGatt = null
                 Logger.debug(TAG, "Disconnected from GATT server.")
+                cleanupCallbacksHandlerThread()
                 resolve("disconnect", "Disconnected.")
             }
         }
@@ -411,7 +431,17 @@ class Device(
         }
         bluetoothGatt?.close()
         connectionState = STATE_CONNECTING
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            initializeCallbacksHandlerThread()
+            bluetoothGatt = device.connectGatt(
+                context,
+                false,
+                gattCallback,
+                BluetoothDevice.TRANSPORT_LE,
+                BluetoothDevice.PHY_OPTION_NO_PREFERRED,
+                callbacksHandler
+            )
+        } else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
             bluetoothGatt = device.connectGatt(
                 context, false, gattCallback, BluetoothDevice.TRANSPORT_LE
             )
@@ -888,6 +918,7 @@ class Device(
             connectionState = STATE_DISCONNECTED
             gatt?.disconnect()
             gatt?.close()
+            cleanupCallbacksHandlerThread()
             reject(key, message)
         }, timeout)
     }